backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
channels==4.0.0
Django==4.2.3
django-cors-headers==4.1.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
